pyproject.toml
[tool.poetry]
name = "cvmaker"
version = "0.1.0"
description = ""
authors = ["Yermakhan <melanhany@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.5"
psycopg2-binary = "^2.9.7"
django-debug-toolbar = "^4.2.0"
black = "^23.9.1"
django-stubs = {extras = ["compatible-mypy"], version = "^4.2.4"}
django-phonenumber-field = "^7.1.0"

[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "cvmaker.settings.dev"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
